tests/requirements.txt
beautifulsoup4~=4.6.0
html5lib~=1.0.1
flask-login~=0.4.1
pytest~=3.6.2
xmlunittest~=0.5.0
